+++ plugins/guests/redhat/cap/box_version.rb
@@ -3,17 +3,19 @@ module VagrantPlugins
     module Cap
       class BoxVersion
         # Prints the version of the vagrant box, parses /etc/os-release for version
-        def self.box_version(machine, script_readable)
+        def self.box_version(machine, options = {})
           command = "cat #{OS_RELEASE_FILE} | grep VARIANT"
+
           # TODO: execute efficient command to solve this
           if machine.communicate.test(command) # test if command is exits with code 0
+            PluginLogger.debug
             machine.communicate.execute(command) do |type, data|
               if type == :stderr
                 @env.ui.error(data)
                 exit 126
               end
 
-              return data.chomp if script_readable
+              return data.chomp if options[:script_readable]
               info = Hash[data.delete('"').split("\n").map { |e| e.split('=') }]
               return "#{info['VARIANT']} #{info['VARIANT_VERSION']}"
             end