+++ plugins/guests/redhat/cap/machine_ip.rb
@@ -6,9 +6,12 @@ module VagrantPlugins
           # Find the guest IP
           command = "ip -o -4 addr show up |egrep -v ': docker|: lo' |tail -1 | awk '{print $4}' |cut -f1 -d\/"
           ip = ''
+
+          PluginLogger.debug
           machine.communicate.execute(command) do |type, data|
             ip << data.chomp if type == :stdout
           end
+
           ip
         end
       end