+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - CHANGELOG fix and README update for OS support for tests @budhrg
 - Fix #188: Name of k8s service not consistent @budhrg
 - Fix #225: service-manager env throws NameError @budhrg
+- Fix #168: Extend --debug flag to show plugin activity @budhrg
 
 ## v1.0.2 May 09, 2016
 - Add --script-readable to env and env docker @bexelbie